+++ CHANGELOG.md
@@ -17,6 +17,8 @@ CHANGELOG
 
 - Update `SummaryEvent` to include the actual name and local file path for locally-executed policy packs.
 
+- Fix Python Dynamic Providers on Windows. [#3855](https://github.com/pulumi/pulumi/pull/3855)
+
 ## 1.9.1 (2020-01-27)
 - Fix a stack reference regression in the Python SDK.
   [#3798](https://github.com/pulumi/pulumi/pull/3798)